Moved part name format validation to InvenTree.validators.py from common.models validation to check if a field exists in part model
